requirements.txt
selenium
#google-chrome wget-install
requests
beautifulsoup4
notion_client
python-dotenv
selenium-stealth
opencv-python
google-cloud-vision
